Flytter omgjørVedtak til egen komponent
