+++ src/frontend/Komponenter/Behandling/Brev/Brev.tsx
@@ -17,11 +17,7 @@ import PdfVisning from './PdfVisning';
 import { ModalWrapper } from '../../../Felles/Modal/ModalWrapper';
 import SystemetLaster from '../../../Felles/SystemetLaster/SystemetLaster';
 import BrevMottakere from '../Brevmottakere/BrevMottakere';
-import {
-    KanIkkeOppretteRevurderingÅrsak,
-    KanOppretteRevurdering,
-} from '../../../App/typer/kanOppretteRevurdering';
-import DataViewer from '../../../Felles/DataViewer/DataViewer';
+import { OmgjørVedtak } from './OmgjørVedtak';
 
 const Brevside = styled.div`
     background-color: var(--navds-semantic-color-canvas-background);
@@ -45,11 +41,6 @@ const AlertStripe = styled(Alert)`
     margin-top: 2rem;
 `;
 
-const AlertContainer = styled.div`
-    padding: 2rem;
-    max-width: 40rem;
-`;
-
 const StyledKnapp = styled(Button)`
     margin-top: 2rem;
 `;
@@ -59,105 +50,6 @@ type Utfall = 'IKKE_SATT' | 'LAG_BREV' | 'OMGJØR_VEDTAK';
 interface IBrev {
     behandlingId: string;
 }
-
-const KanOppretteRevurderingTekst: React.FC<{ kanOppretteRevurdering: KanOppretteRevurdering }> = ({
-    kanOppretteRevurdering,
-}) => {
-    if (kanOppretteRevurdering.kanOpprettes) {
-        return (
-            <Alert variant={'info'}>
-                Resultatet av klagebehandlingen er at påklaget vedtak skal omgjøres. Du kan nå
-                ferdigstille klagebehandlingen og en revurderingsbehandling for å fatte nytt vedtak
-                blir automatisk opprettet.
-            </Alert>
-        );
-    } else if (kanOppretteRevurdering.årsak === KanIkkeOppretteRevurderingÅrsak.ÅPEN_BEHANDLING) {
-        return (
-            <Alert variant={'warning'}>
-                Resultatet av klagebehandlingen er at påklaget vedtak skal omgjøres. En
-                revurderingsbehandling for å fatte nytt vedtak blir ikke automatisk opprettet da
-                bruker allerede har en åpen behandling.
-            </Alert>
-        );
-    } else {
-        return (
-            <Alert variant={'warning'}>
-                Resultatet av klagebehandlingen er at påklaget vedtak skal omgjøres. En
-                revurderingsbehandling for å fatte nytt vedtak blir ikke automatisk opprettet.
-            </Alert>
-        );
-    }
-};
-
-export const OmgjørVedtak: React.FC<{
-    behandlingId: string;
-    ferdigstill: () => void;
-    senderInn: boolean;
-}> = ({ behandlingId, ferdigstill, senderInn }) => {
-    const { axiosRequest } = useApp();
-    const { behandlingErRedigerbar } = useBehandling();
-    const [visModal, settVisModal] = useState<boolean>(false);
-    const [feilmelding, settFeilmelding] = useState('');
-    const [kanOppretteRevurdering, settKanOppretteRevurdering] = useState<
-        Ressurs<KanOppretteRevurdering>
-    >(byggTomRessurs());
-
-    const lukkModal = () => {
-        settVisModal(false);
-        settFeilmelding('');
-    };
-
-    useEffect(() => {
-        if (behandlingErRedigerbar) {
-            axiosRequest<KanOppretteRevurdering, null>({
-                method: 'GET',
-                url: `/familie-klage/api/behandling/${behandlingId}/kan-opprette-revurdering`,
-            }).then(settKanOppretteRevurdering);
-        }
-    }, [axiosRequest, behandlingErRedigerbar, behandlingId]);
-
-    return (
-        <DataViewer response={{ kanOppretteRevurdering }}>
-            {({ kanOppretteRevurdering }) => (
-                <>
-                    {behandlingErRedigerbar && (
-                        <AlertContainer>
-                            <KanOppretteRevurderingTekst
-                                kanOppretteRevurdering={kanOppretteRevurdering}
-                            />
-                            <StyledKnapp onClick={() => settVisModal(true)}>
-                                Ferdigstill
-                            </StyledKnapp>
-                        </AlertContainer>
-                    )}
-                    {!behandlingErRedigerbar && (
-                        <AlertContainer>
-                            <Alert variant={'info'}>
-                                Brev finnes ikke fordi klagen er tatt til følge.
-                            </Alert>
-                        </AlertContainer>
-                    )}
-                    <ModalWrapper
-                        tittel={'Bekreft ferdigstillelse av klagebehandling'}
-                        visModal={visModal}
-                        onClose={() => lukkModal()}
-                        aksjonsknapper={{
-                            hovedKnapp: {
-                                onClick: ferdigstill,
-                                tekst: 'Ferdigstill',
-                                disabled: senderInn,
-                            },
-                            lukkKnapp: { onClick: lukkModal, tekst: 'Avbryt' },
-                            marginTop: 4,
-                        }}
-                    >
-                        {feilmelding && <AlertStripe variant={'error'}>{feilmelding}</AlertStripe>}
-                    </ModalWrapper>
-                </>
-            )}
-        </DataViewer>
-    );
-};
 
 export const Brev: React.FC<IBrev> = ({ behandlingId }) => {
     const [brevRessurs, settBrevRessurs] = useState<Ressurs<string>>(byggTomRessurs());
